Cloud Functions/mysql-to-gcs/requirements.txt
# package>=version
#python-csv
mysql-connector-python
google-cloud-storage
